Debug "swooshing" movement
Fixes bug in commit 78d6041, where a piece in the right-most column would "swoosh" to the left-most column after trying to move to the right
Bug was caused by the x-value in the piece notation changing from 9 to 10, which caused the return_x_y function to recognize it as 1 (0 being part of the y-value)
Now, the move_piece_right/left/down functions check if the piece is too far to the left/right/bottom, and returns None if it does.
This might become a problem later because the piece notation cannot handle two-digit x values.